♻️ refactor(monitoring): extract codex inspection storage
Move Codex inspection last-run serialization, hydration, cache IO, and result sorting into a dedicated storage model.

The service keeps its public exports through re-exports while execution logic stays unchanged; validated with type-check, codex inspection tests, and lint.
